Move the N module out to its own N.d.ts and move EntryPoints to N/types
