+++ N/types.d.ts
@@ -1,69 +1,8 @@
-import * as N_auth from './auth';
-import * as N_config from './config';
-import * as N_crypto from './crypto';
-import * as N_currency from './currency';
-import * as N_currentRecord from './currentRecord';
-import * as N_email from './email';
-import * as N_encode from './encode';
-import * as N_error from './error';
-import * as N_file from './file';
-import * as N_format from './format';
 import * as N_http from './http';
-import * as N_https from './https';
-import * as N_log from './log';
-import * as N_plugin from './plugin';
 import * as N_portlet from './portlet';
 import * as N_record from './record';
-import * as N_redirect from './redirect';
-import * as N_render from './render';
-import * as N_runtime from './runtime';
 import * as N_search from './search';
-import * as N_sso from './sso';
-import * as N_task from './task';
-import * as N_transaction from './transaction';
-import * as N_url from './url';
-import * as N_util from './util';
-import * as N_workflow from './workflow';
-import * as N_xml from './xml';
-import * as N_ui_dialog from './ui/dialog';
-import * as N_ui_message from './ui/message';
 import * as N_ui_serverWidget from './ui/serverWidget';
-
-export {N_auth as auth};
-export {N_config as config};
-export {N_crypto as crypto};
-export {N_currency as currency};
-export {N_currentRecord as currentRecord};
-export {N_email as email};
-export {N_encode as encode};
-export {N_error as error};
-export {N_file as file};
-export {N_format as format};
-export {N_http as http};
-export {N_https as https};
-export {N_log as log};
-export {N_plugin as plugin};
-export {N_portlet as portlet};
-export {N_record as record};
-export {N_redirect as redirect};
-export {N_render as render};
-export {N_runtime as runtime};
-export {N_search as search};
-export {N_sso as sso};
-export {N_task as task};
-export {N_transaction as transaction};
-export {N_url as url};
-export {N_util as util};
-export {N_workflow as workflow};
-export {N_xml as xml};
-
-declare interface N_UI_Module {
-    dialog: typeof N_ui_dialog,
-    message: typeof N_ui_message,
-    serverWidget: typeof N_ui_serverWidget,
-}
-declare var N_ui: N_UI_Module;
-export {N_ui as ui};
 
 /*Don't export these into the Namespace as we don't
 want to accidentally use a comparison like this:
