+++ tsconfig.json
@@ -10,8 +10,7 @@
     "baseUrl": ".",
     "lib": ["es5", "es2015.promise", "dom"],
     "paths": {
-      "N": ["node_modules/@hitc/netsuite-types/N"],
-      "N/*": ["node_modules/@hitc/netsuite-types/N/*"]
+      "N/*": ["N/*"]
     }
   },
   "exclude": ["node_modules"]
